Allow to override default service and app deployment services
So developers can use custom implementation of deployers, in case if any
customization of apps or services list is required.
