+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/deployer/BackingAppDeploymentService.java
@@ -19,54 +19,12 @@ package org.springframework.cloud.appbroker.deployer;
 import java.util.List;
 
 import reactor.core.publisher.Flux;
-import reactor.core.scheduler.Schedulers;
-import reactor.util.Logger;
-import reactor.util.Loggers;
 
-public class BackingAppDeploymentService {
-	private final Logger log = Loggers.getLogger(BackingAppDeploymentService.class);
+public interface BackingAppDeploymentService {
 
-	private final DeployerClient deployerClient;
+	Flux<String> deploy(List<BackingApplication> backingApps, String serviceInstanceId);
 
-	public BackingAppDeploymentService(DeployerClient deployerClient) {
-		this.deployerClient = deployerClient;
-	}
+	Flux<String> update(List<BackingApplication> backingApps, String serviceInstanceId);
 
-	public Flux<String> deploy(List<BackingApplication> backingApps, String serviceInstanceId) {
-		return Flux.fromIterable(backingApps)
-			.parallel()
-			.runOn(Schedulers.parallel())
-			.flatMap(backingApplication -> deployerClient.deploy(backingApplication, serviceInstanceId))
-			.sequential()
-			.doOnRequest(l -> log.debug("Deploying applications {}", backingApps))
-			.doOnEach(response -> log.debug("Finished deploying application {}", response))
-			.doOnComplete(() -> log.debug("Finished deploying application {}", backingApps))
-			.doOnError(exception -> log.error("Error deploying applications {} with error '{}'",
-				backingApps, exception.getMessage()));
-	}
-
-	public Flux<String> update(List<BackingApplication> backingApps, String serviceInstanceId) {
-		return Flux.fromIterable(backingApps)
-			.parallel()
-			.runOn(Schedulers.parallel())
-			.flatMap(backingApplication -> deployerClient.update(backingApplication, serviceInstanceId))
-			.sequential()
-			.doOnRequest(l -> log.debug("Updating applications {}", backingApps))
-			.doOnEach(response -> log.debug("Finished updating application {}", response))
-			.doOnComplete(() -> log.debug("Finished updating application {}", backingApps))
-			.doOnError(exception -> log.error("Error updating applications {} with error {}", backingApps, exception));
-	}
-
-	public Flux<String> undeploy(List<BackingApplication> backingApps) {
-		return Flux.fromIterable(backingApps)
-			.parallel()
-			.runOn(Schedulers.parallel())
-			.flatMap(deployerClient::undeploy)
-			.sequential()
-			.doOnRequest(l -> log.debug("Undeploying applications {}", backingApps))
-			.doOnEach(response -> log.debug("Finished undeploying application {}", response))
-			.doOnComplete(() -> log.debug("Finished undeploying application {}", backingApps))
-			.doOnError(exception -> log.error("Error undeploying applications {} with error '{}'",
-				backingApps, exception.getMessage()));
-	}
+	Flux<String> undeploy(List<BackingApplication> backingApps);
 }
